refactor(core): move extraction parser into workflow
